pyproject.toml
[project]
name = "aquamonitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=2.0.2",
    "pillow>=11.1.0",
    "tqdm>=4.67.1",
]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "datasets>=3.3.1",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "pytest>=8.3.4",
]
